add slice() function to some twig template
