+++ composer.json
@@ -17,7 +17,6 @@
         "ramsey/uuid": "^4.1",
         "phpmailer/phpmailer": "^6.4",
         "tinymce/tinymce": "^5.8",
-        "cocur/slugify": "^4.0",
-        "twig/string-extra": "^3.3"
+        "cocur/slugify": "^4.0"
     }
 }
