+++ src/Controller/admin/ContactController.php
@@ -4,7 +4,6 @@ namespace App\Controller\admin;
 
 use App\Core\BaseController;
 use App\Repository\Manager\ContactManager;
-use App\Services\DateFormat;
 use App\Services\PHPSession;
 use Ramsey\Uuid\Uuid;
