+++ src/templates/viewComment.html.twig
@@ -1,0 +1,47 @@
+{% extends "base.html.twig" %}
+
+{% block metatitle %}
+Commentaire posté par {{ comment.pseudo }}, {{ comment.date }} | Blog de Floryss Rubechi
+{% endblock %}
+
+{% block metadescription %}
+Partie administration : page individuelle de commentaire avec le contenu non-tronqué et possibilité de le valider.
+{% endblock %}
+
+{% block title %}
+Voir un commentaire
+{% endblock %}
+
+{% block css %}
+<link type="text/css" rel="stylesheet" href="https://cdn.datatables.net/1.10.24/css/jquery.dataTables.min.css">
+{% endblock %}
+
+
+
+{% block body %}
+<div class="maincontent xlargecontainer">
+    <section>
+        <h1>Commentaire posté par {{ comment.pseudo }}<br>{{ comment.date }}</h1>
+        <div class="textCenter">
+            <p class="marginBot50 italic"> " {{ comment.content }} "</p>
+            {% if admin %}
+                {% if comment.isValidated %}
+                    <p class="textGreen">Ce commentaire est accepté.</p>
+                    <p><a href="invalider-commentaire?idComment={{ comment.id }}" onclick="return confirm('Etes vous sûr d invalider ce commentaire ?')">Invalider ce commentaire</a></p>
+                {% else %}
+                    <p class="textRed">Ce commentaire n'est pas accepté.</p>
+                    <p><a href="valider-commentaire?idComment={{ comment.id }}" onclick="return confirm('Etes vous sûr de valider ce commentaire ?')">Valider ce commentaire</a></p>
+                {% endif %}
+                <p><a href="/blogphp/admin-commentaires">Revenir à la liste des commentaires</a></p>
+            {% else %}
+                {% if comment.isValidated %}
+                    <p class="textGreen">Ce commentaire est accepté.</p>
+                {% else %}
+                    <p class="textRed">Ce commentaire n'est pas accepté.</p>
+                {% endif %}
+                <p><a href="dashboard">Retourner au dashboard</a></p>
+            {% endif %}
+        </div>
+    </section>
+</div>
+{% endblock %}
